Rename .gitignore to PromoteApp.m
This class was created to promote one of my apps. I wanted to show the user different messages and capture responses to know which message resulted in high response rate. The class is configurable and ready to be used in other apps as well. There are some code outside this class that make it work though.